src/krishmodel/backend/requirements.txt
nibabel==5.2.0
numpy==1.24.3
Pillow==10.2.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
